Adicionando poc da search bar
