+++ Estudos/SearchBar/App/General/Delegate/System/AppDelegate.swift
