+++ Estudos/SearchBar/App/Screens/Search/DataSource/NameCollectionDataSource.swift
