+++ Estudos/SearchBar/App/Screens/Search/View/Cell/NameCell.swift
