--- Estudos/SearchBar/POC-Search-Seg-Tab/General/Delegate/System/SceneDelegate.swift
@@ -1,26 +1,0 @@
-//
-//  SceneDelegate.swift
-//  POC-Search-Seg-Tab
-//
-//  Created by Gui Reis on 09/09/22.
-//
-
-import UIKit
-
-class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-
-    var window: UIWindow?
-
-
-    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
-        guard let windowScene = (scene as? UIWindowScene) else { return }
-        
-        self.window = UIWindow(windowScene: windowScene)
-        self.window?.rootViewController = MainController()
-        self.window?.makeKeyAndVisible()
-    }
-
-    func sceneDidEnterBackground(_ scene: UIScene) {
-        (UIApplication.shared.delegate as? AppDelegate)?.saveContext()
-    }
-}
