--- Estudos/SearchBar/Shared/Assets/Assets.xcassets/Colors App/Buttons/Contents.json
@@ -1,6 +1,0 @@
-{
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
