--- Estudos/SearchBar/Shared/Assets/Assets.xcassets/Colors App/Buttons/Pink_ED7070.colorset/Contents.json
@@ -1,38 +1,0 @@
-{
-  "colors" : [
-    {
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "0.439",
-          "green" : "0.439",
-          "red" : "0.929"
-        }
-      },
-      "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "0.439",
-          "green" : "0.439",
-          "red" : "0.929"
-        }
-      },
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
