+++ Estudos/SearchBar/Shared/Core Data/POC_Search_Seg_Tab.xcdatamodeld/.xccurrentversion
@@ -1,8 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
-<dict>
-	<key>_XCCurrentVersionName</key>
-	<string>POC_Search_Seg_Tab.xcdatamodel</string>
-</dict>
+<dict/>
 </plist>
