--- Estudos/SearchBar/Shared/Desgin System/Adaptive Layout/Model/Enum/Dimension.swift
@@ -1,12 +1,0 @@
-/* Macro - Grupo 05 */
-
-
-/// Tipos de dimensão de uma tela
-public enum Dimension {
-    
-    /// Largura
-    case width
-    
-    /// Altura
-    case height
-}
