--- Estudos/SearchBar/Shared/Desgin System/Models/Icons/Extension/UIImage+AppIcons.swift
@@ -1,12 +1,0 @@
-/* Macro - Grupo 05 */
-
-/* Bibliotecas necessárias: */
-import UIKit
-
-
-extension UIImage {
-    
-    convenience init?(_ icon: AppIcons) {
-        self.init(systemName: icon.description)
-    }
-}
