--- Estudos/SearchBar/Shared/Desgin System/Models/Icons/IconInfo.swift
@@ -1,13 +1,0 @@
-/* Macro - Grupo 05 */
-
-/* Bibliotecas necessárias: */
-import UIKit
-
-
-/// Configurações do ícone de um botão (UIButton)
-struct IconInfo{
-    let icon: AppIcons
-    let size: CGFloat
-    let weight: UIImage.SymbolWeight
-    let scale: UIImage.SymbolScale
-}
