+++ Estudos/SearchBar/Shared/Desgin System/Views/CollectionGroup/CollectionGroupStyle.swift
